// applets/sonarr/api/model/seasonResource.ts
/**
 * Generated by orval v6.11.1 🍺
 * Do not edit manually.
 * Sonarr
 * Sonarr API docs
 * OpenAPI spec version: 3.0.0
 */
import type { SeasonStatisticsResource } from "./seasonStatisticsResource";
import type { MediaCover } from "./mediaCover";

export interface SeasonResource {
  seasonNumber?: number;
  monitored?: boolean;
  statistics?: SeasonStatisticsResource;
  images?: MediaCover[] | null;
}